+++ code/src/nuvla/connector/nuvlabox_connector.py
@@ -3,8 +3,8 @@
 import requests
 import logging
 from .connector import Connector, should_connect
-from .utils import create_tmp_file, utc_from_now_iso, unique_id
-from ..job.actions.nuvla import Notification
+from .utils import create_tmp_file
+
 
 class NuvlaBoxConnector(Connector):
     def __init__(self, **kwargs):
@@ -15,14 +15,13 @@ class NuvlaBoxConnector(Connector):
         self.ssl_file = None
         self.nuvlabox_api = requests.Session()
         self.nuvlabox_api.verify = False
-        self.nuvlabox_api.headers = {'Content-Type': 'application/json', 'Accept': 'application/json'}
+        self.nuvlabox_api.headers = {'Content-Type': 'application/json',
+                                     'Accept': 'application/json'}
 
         self.nuvlabox_id = kwargs.get("nuvlabox_id")
         self.nuvlabox = None
         self.timeout = 60
         self.acl = None
-
-        self.notification = Notification(self.api)
 
     @property
     def connector_type(self):
@@ -52,8 +51,9 @@ class NuvlaBoxConnector(Connector):
                 infra_service = self.api.get(service_id).data
                 if infra_service.get("subtype") == 'swarm':
                     credentials = self.api.search('credential',
-                                                  filter='parent="{}" and subtype="{}"'.format(service_id,
-                                                                                               cred_subtype)).resources
+                                                  filter='parent="{}" and subtype="{}"'.format(
+                                                      service_id,
+                                                      cred_subtype)).resources
 
                     return credentials[0].data
 
@@ -62,7 +62,8 @@ class NuvlaBoxConnector(Connector):
         try:
             secret = credential['cert'] + '\n' + credential['key']
         except KeyError:
-            logging.error("Credential for {} is either missing or incomplete".format(self.nuvlabox.get("id")))
+            logging.error(
+                "Credential for {} is either missing or incomplete".format(self.nuvlabox.get("id")))
             raise
 
         self.ssl_file = create_tmp_file(secret)
@@ -101,40 +102,27 @@ class NuvlaBoxConnector(Connector):
         if nb_api_endpoint:
             self.job.set_progress(50)
         else:
-            logging.warning("NuvlaBox {} missing API endpoint in its status resource".format(self.nuvlabox.get("id")))
-            raise("NuvlaBox {} missing API endpoint in its status resource".format(self.nuvlabox.get("id")))
+            logging.warning("NuvlaBox {} missing API endpoint in its status resource".format(
+                self.nuvlabox.get("id")))
+            raise ("NuvlaBox {} missing API endpoint in its status resource".format(
+                self.nuvlabox.get("id")))
 
         # 2nd - get the corresponding credential and prepare the SSL environment
         self.setup_ssl_credentials()
 
         self.job.set_progress(90)
-        action_endpoint = '{}/{}'.format(nb_api_endpoint, kwargs.get('api_action_name', '')).rstrip('/')
+        action_endpoint = '{}/{}'.format(nb_api_endpoint,
+                                         kwargs.get('api_action_name', '')).rstrip('/')
 
         method = kwargs.get('method', 'GET').upper()
         payload = kwargs.get('payload', {})
 
         # 3rd - make the request
-        r = self.nuvlabox_api.request(method, action_endpoint, json=payload, timeout=self.timeout).json()
-        self.job.set_progress(99)
-
-        msg = 'Call /api/{} for NuvlaBox {}. Output: {}'.format(kwargs.get('api_action_name', ''),
-                                                                self.nuvlabox_id,
-                                                                r)
-
-        self.create_notification(msg)
+        r = self.nuvlabox_api.request(method, action_endpoint, json=payload,
+                                      timeout=self.timeout).json()
+        self.job.set_progress(100)
 
         return r
-
-    def create_notification(self, msg):
-
-        expiry = utc_from_now_iso(24 * 3600)
-        notification_unique_id = unique_id(self.nuvlabox_id, expiry)
-
-        notification_id = self.notification.create(msg, 'nuvlabox-action', notification_unique_id, expiry=expiry,
-                                                   target_resource=self.nuvlabox_id, acl=self.acl)
-
-        logging.info('Created notification {0}'.format(
-            notification_id))
 
     @should_connect
     def stop(self, **kwargs):