+++ code/src/nuvla/job/executor.py
@@ -69,7 +69,7 @@ class Executor(Base):
                 status_message = type(ex).__name__ + '-' + ''.join(traceback.format_exception(
                     etype=ex_type, value=ex_msg, tb=ex_tb))
                 logging.error('Failed to process {}, with error: {}'.format(job.id, status_message))
-                job.update_job(state='FAILED', status_message=status_message)
+                job.update_job(state='FAILED', status_message=status_message, return_code=1)
             else:
                 state = 'SUCCESS' if return_code == 0 else 'FAILED'
                 job.update_job(state=state, return_code=return_code)